chore: rework QueryPipeline API to take a generic qbvh updater
This allows to pass less parameters depending on the updating mode.
