+++ CHANGELOG.md
@@ -53,6 +53,10 @@ This release introduces two new crates:
   `bevy_rapier` plumbings, but it is no longer useful. Adding a collider must always go througthe `ColliderSet`.
 - `CharacterController::solve_character_collision_impulses` now takes multiple `CharacterCollision` as parameter:
   this change will allow further internal optimizations.
+- `QueryPipeline::update` now doesn't need the `RigidBodySet` as parameter.
+- Removed `QueryPipelineMode`.
+- `QueryPipeline::update_with_mode` now takes `impl QbvhDataGenerator<ColliderHandle>` as parameter.
+  see [`QueryPipeline::updaters`] module for more information.
 
 ## v0.19.0 (05 May 2024)
